requirements.txt
gensim==3.8.1
numpy==1.21.0
matplotlib
scikit-learn
jieba 
tensorflow-gpu==2.10.0
protobuf==3.19.6 -i https://pypi.tuna.tsinghua.edu.cn/simple
PySimpleGUI
torch==1.13.1
torchvision==0.14.1 
torchtext==0.14.1 
seaborn